+++ src/Command/CreateUserCommand.php
@@ -75,9 +75,9 @@ final class CreateUserCommand extends AbstractUserCommand
         } catch (ValidationFailedException $ex) {
             $io->validationError($ex);
 
-            return (int) Command::FAILURE;
+            return Command::FAILURE;
         }
 
-        return (int) Command::SUCCESS;
+        return Command::SUCCESS;
     }
 }
