+++ src/Command/ReloadCommand.php
@@ -118,7 +118,7 @@ final class ReloadCommand extends Command
             sprintf('Kimai config was reloaded')
         );
 
-        return (int) Command::SUCCESS;
+        return Command::SUCCESS;
     }
 
     private function rebuildCaches(string $environment, SymfonyStyle $io, InputInterface $input, OutputInterface $output)
